requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
google-genai==1.12.1
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3

